+++ packages/pattern_dots/CHANGELOG.md
@@ -1,3 +1,4 @@
+## 1.0.0+1 - 2024-04-16
 ## 1.0.0 - 2024-02-04
 ### Changed
 - rewrite whole apis, simpplify api usage