+++ packages/webview_refresher/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2 - 2024-04-19
+### Added
+- add platform prop
+
+### Fixed
+- fix a scroll issue
+- fix `example` refresh cant finish
+- remove unused print
+
 ## 0.0.1+1 - 2024-04-16
 ### Changed
 - update README.md